@@ -14,19 +14,19 @@
     "@fortawesome/free-brands-svg-icons": "^6.4.2",
     "@fortawesome/free-solid-svg-icons": "^6.4.2",
     "@fortawesome/react-fontawesome": "^0.2.0",
-    "@types/node": "20.2.5",
-    "@types/react": "18.2.7",
-    "@types/react-dom": "18.2.4",
-    "eslint": "8.41.0",
-    "eslint-config-next": "13.4.4",
+    "@types/node": "20.8.6",
+    "@types/react": "18.2.28",
+    "@types/react-dom": "18.2.13",
+    "eslint": "8.51.0",
+    "eslint-config-next": "13.5.5",
     "fs": "0.0.1-security",
     "isotope-layout": "^3.0.6",
-    "next": "13.4.4",
+    "next": "13.5.5",
     "next-pwa": "^5.6.0",
     "path": "^0.12.7",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "typed.js": "^2.0.16",
-    "typescript": "5.0.4"
+    "typescript": "5.2.2"
   }
 }
